requirement.txt

pyzmq
nicegui
pandas